split rocket status into 2 components
