+++ web/src/lib/common/layoutStore.ts
@@ -13,6 +13,7 @@ import LogViewer from '$lib/components/LogViewer.svelte';
 import Settings from '$lib/components/Settings.svelte';
 import RocketTracker from '$lib/components/smart/RocketTracker.svelte';
 import RocketStatus from '$lib/components/RocketStatus.svelte';
+import RocketImu from '$lib/components/RocketIMU.svelte';
 import { browser } from '$app/environment';
 
 export const layoutComponents = {
@@ -27,6 +28,7 @@ export const layoutComponents = {
 	GenericSbgGraph,
 	LayoutList,
 	RocketStatus,
+	RocketImu
 } as const;
 
 export const layoutComponentsString = Object.keys(layoutComponents);
